Always go to state write_and_go after conn_mwrite
There was a mysterious todo comment by 'XXX', that tested
whether protocol was binary, and only uses write_and_go
if the protocol was binary.

Instead, I looked for everywhere where we set the conn state
to conn_mwrite, and also setup the write_and_go state to
conn_new_cmd, so we don't need a special case for
binary_prot after the transmit().