Added upstream_suffix to allow final END to multi-get
Previously, the client would see multiple END statements,
because each downstream conn would produce another END.
Instead, we use the downstream END to tell us that
a downstream conn is done and should be paused.

The pausing decrements a downstream_used counter.
When that counter hits 0, then the downstream is
released.  During the release, we see if there's any
upstream_suffix that needs to be written to the
upstream_conn.