Changes to create tighter binding to libconflate and libmemcached
The way libconflate was being built previously was incorrect as it
was assuming it as a system module.  This change brings it in under
the overall proxybuild.  It may later move to a subdir from moxi.

libmemcached is now a subdir from moxi (checked out from
lp:libmemcached) and uses libtool to link in.