fixed running of t/moxi_all.pl under dash shell
The problem was caused by dash inability to use exec when launching
./moxi-debug for tests. This caused $childpid to be dash process, not
moxi-debug process. This in turn caused SIGINT for ./moxi-debug to be
directed to dash, which left ./moxi-debug process alife. The use of
backtick operator in ./moxi.pl to read test status and inheritance of
stdin & stdout to ./moxi-debug caused backtick operator to wait
forever, 'cause ./moxi-debug still had stdout open.

I took approach that I saw in check unit testing framework. We simply
detach child process to it's own process group and send SIGINT to whole
group.

The only possible downside is mysterious hangs that'll happen
when ./moxi-debug inherits real terminal as it's stdout and tries to
print anything there and TOSTOP terminal attribute is set (it seems to
be off by default on all platforms).