Updated autogen.sh to recursively clean git submodules. Helps with CI.
When using buildbot, we should be building from a clean rev of the
submodule.  Previous builds may leave droppings which makes for a
less realistic build.  Wanting continuious integration to be as close
to what a user would do, we now clean all of the submodules at autogen.sh
time.  This will make slightly longer CI builds, but much more correct.