Updated use of c99 for Intel ICC.  Typedef required to get u_char.
It is not 100% clear what is happening here, but from research it
seems that gcc frequently defines _GNU_SOURCE which then tries to get
libc to conform to those standards.  Various resources seem to
indicate that _GNU_SOURCE is not always defined by gcc, some distros
have their gcc configured differently.

libevent has a u_char in it's header file, and even has a typedef
for it.  icc seems to ignore this typedef (which may actually be
the right thing to do!).  By re-adding this typedef only, the libevent
test will pass and the later compilation using regular includes of
event.h behave correctly.

It should be noted that u_char does not appear in the C99 standard.

Resources:
http://www.gnu.org/s/libc/manual/html_node/Feature-Test-Macros.html
http://blogs.sun.com/rvs/entry/fun_with_libc_compile_time