cproxy_binary_uncork_cmds()
In this change, the uncorking of quiet binary commands is finally
implemented, with a proper refcounting of item structs.

b2b_forward_item() and b2b_forward_item_vbucket() helper functions
were refactored into existence to allow the uncorking code to easily
"replay" all the corked commands.

Along the way, having a vbucketId of -1 turned out to be meaningful in
this commit where it means we really don't have a vbucketId, like a
SQL NULL.  vbucketId of -1 occurs in the code when we're doing a
non-key'ed broadcast command, like a NOOP or FLUSH_ALL.

I also discovered we shouldn't clean up the corked commands
linked-list during conn_cleanup(), but instead do the cleanup on
conn_free().  It turns out conn_cleanup() is invoked every time
through the drive_machine() loop when we move back to conn_new_cmd
state, which made the code incorrectly forget all the corked quiet
commands.

Change-Id: I0d371a3a0c964e01cd4579de02c7244106275479
Reviewed-on: http://review.northscale.com/944
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Steve Yen <steve.yen@gmail.com>