moxi tests pass again
With this change, the old moxi 'good tests' inherited from memcached
pass again, with the testing topologies updated for libvbucket (the
new default hashlib).  The change still allows the previous
libmemcached hashlib as an option.

To run the tests, use...

  ./t/moxi_all.pl

And that's also part of 'make test' now, with better exit code
propagation.

Change-Id: Ie9bdd82170825f731efaa41ef1624babfcd17743
Reviewed-on: http://review.northscale.com/1162
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>