bug 1792 - handle OPAQUE_IGNORE_REPLY for both ascii and binary clients
This commit is with regards to handling ascii and binary clients concurrently.

A client on one protocol (binary), might be assigned to use a
downstream connection that was previously servicing another protocol
(ascii).  The a2b protocol handler would use a special opaque magic
value OPAQUE_IGNORE_REPLY to signal that a given binary response
should be ignored.  This OPAQUE_IGNORE_REPLAY was used to map ascii
"noreply" commands to binary quiet commands.

This commit allows the b2b (binary-to-binary) codepath to handle any
OPAQUE_IGNORE_REPLY which it encounters.  These responses might be
sitting on the connection due to a2b (ascii-to-binary) usage.

cproxy_binary_ignore_reply()

Change-Id: Ie0232c3c72268d352d2ed553283b2abab6db7874
Reviewed-on: http://review.northscale.com/1229
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>