bug 1497 - binary-2-binary broadcast command impl was wrong
Before, the code didn't correctly handle a cluster situation for
broadcast commands and stats merging was wrong for binary-to-binary
commands.

With this commit, cluster-wide binary-to-binary broadcast commands
(like NO-OP, FLUSH, and STATS) utilize the same mechanism like
ascii-2-binary broadcast commands.

That is, there is an upstream_suffix set of bytes associated with each
client broadcast command.  This upstream_suffix is the last response
message that is sent to back to the upstream client after all replies
have been gathered back from the downstream servers.

In the binary-to-binary case, this upstream_suffix would be a NO-OP,
FLUSH, or STATS response.  (In the ascii, case, the upstream_suffix
would be something like "END", or "OK").

The upstream_suffix was extended to also have an optional, explicit
upstream_suffix_len, which the code will use in the binary protocol
case instead of doing a strlen(upstream_suffix).

Finally, binary-to-binary stats was not doing the stats merging at all.
That should admittedly be a separate commit.

Change-Id: I9d105aa402c7e9f2c930aa3f51226ee7d569b539
Reviewed-on: http://review.northscale.com/1256
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Steve Yen <steve.yen@gmail.com>