notes on zstored-conn-pooling semantics for cproxy_connect_downstream()
Added some commens/notes on a change to the return value of
cproxy_connect_downstream(), where it can now return -1 in the
case that a downstream connection isn't ready/assigned yet (but
not a full error either).  In that case, the upstream connection
should remain in the conn_pause state, and the upstream connection
should be successfully assigned to a downstream struct.

The downstream struct, in turn, on a -1 return value case, will in
future code be enqueued to wait for an assigned downstream conn.

A downstream struct represents a set of connections.

Change-Id: I8f3b097fc0a63023c6fe92c1219b729c69565a02
Reviewed-on: http://review.northscale.com/2241
Reviewed-by: Manik Taneja <mtaneja@zynga.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>