mcmux protocol support for moxi
Added support for ASCII protocol extension to moxi. To enable
mcmux compatiblity -X option should be passed as one of the command
line paramters along with the port number that moxi should listen on.

./moxi -p 11311 -X

when started in this mode, moxi expects the mcmux protocol string
to be appended before the start of a memcache command.
An example of an interaction between client and moxi :

A:localhost:11411 get k1
VALUE k1 0 5
manik
END

get k1
ERROR

Amended version: check for null d->upstream_conn
Use a stack variable for msst instead of allocating it.

Change-Id: I40d853023c80d08a40113e338e5bb6708e7f4bd9
Reviewed-on: http://review.northscale.com/2317
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>